@@ -1,7 +1,8 @@
 package com.lwg.commons.adapter;
 
-import android.support.v7.widget.RecyclerView;
+
 import android.view.ViewGroup;
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * Interface to create and bind {@link RecyclerView.ViewHolder}
