@@ -1,10 +1,12 @@
 package com.lwg.commons.adapter;
 
-import android.support.annotation.LayoutRes;
-import android.support.v7.widget.RecyclerView;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.annotation.LayoutRes;
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * Generic delegate adapter that allows to inflate views inside the recycler view with no dynamic binding
